Fixed dtype issue when checking for ring geometry

+++ holoviews/core/data/multipath.py
@@ -551,7 +551,8 @@ def ensure_ring(geom, values=None):
     """
     if values is None:
         values = geom
-    breaks = np.where(np.isnan(geom).sum(axis=1))[0]
+    
+    breaks = np.where(np.isnan(geom.astype('float')).sum(axis=1))[0]
     starts = [0] + list(breaks+1)
     ends = list(breaks-1) + [len(geom)-1]
     zipped = zip(geom[starts], geom[ends], ends, values[starts])
